# rate: reduce allocation and speed up ReserveN