benchmark               old ns/op     new ns/op     delta
BenchmarkReserveN-4     120           81.9          -31.75%

benchmark               old allocs     new allocs     delta
BenchmarkReserveN-4     1              0              -100.00%

benchmark               old bytes     new bytes     delta
BenchmarkReserveN-4     64            0             -100.00%