@kulti 2019-06-24
Fixed in go 1.12